requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
python-dotenv==1.0.1
reportlab==4.2.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
qrcode[pil]==7.4.2
Pillow==10.3.0
Flask-Mail==0.10.0
